test(settings): split app preference coverage
Why:
- Continue moving package-owned AppAndReader tests out of the app-host bundle.
- Keep preference registry, reset, and settings-search contracts in the package lanes that own the behavior.

What Changed:
- Added AppPreferenceRegistryTests to BibleCoreTests for registry, default, normalizer, action-row, reset-key, and reset behavior contracts.
- Added SettingsSearchMatcherTests to BibleUITests for settings-search presentation matching.
- Removed the migrated tests from AndBibleTests+AppAndReader.
- Updated the migration plan ledger for the AppAndReader settings slice.
- Tightened the reset test to snapshot and restore touched UserDefaults keys.

Validation:
- git diff --cached --check.
- python3 scripts/check_repo_standards.py docblocks --all-files.
- python3 scripts/check_repo_standards.py source-guards.
- swift package describe --type json.
- xcodebuild test BibleCoreTests/AppPreferenceRegistryTests on iPhone 16 iOS 18.6.
- xcodebuild test BibleUITests/SettingsSearchMatcherTests on iPhone 16 iOS 18.6.
- npx gitnexus analyze.
- GitNexus staged detect_changes reported LOW risk with no affected execution flows.

Impact:
- Reduces app-host unit-test surface with no production behavior changes.
- Preserves eight migrated test contracts in package targets.
- Improves package-test isolation for UserDefaults-backed preferences.
