+++ AndBibleTests/AndBibleTests+AppAndReader.swift
@@ -19,170 +19,6 @@ import struct SwiftUI.Color
 #endif
 
 extension AndBibleTests {
-    func testAppPreferenceRegistryHasDefinitionForAllKeys() {
-        let keys = AppPreferenceKey.allCases
-        XCTAssertEqual(keys.count, 40)
-        XCTAssertEqual(Set(keys).count, keys.count)
-        XCTAssertEqual(AppPreferenceRegistry.definitions.count, keys.count)
-
-        for key in keys {
-            XCTAssertEqual(AppPreferenceRegistry.definition(for: key).key, key)
-        }
-    }
-
-    func testCriticalPreferenceDefaultsMatchParityContract() {
-        XCTAssertEqual(AppPreferenceRegistry.stringDefault(for: .nightModePref3), "system")
-        XCTAssertEqual(AppPreferenceRegistry.stringDefault(for: .toolbarButtonActions), "default")
-        XCTAssertEqual(AppPreferenceRegistry.stringDefault(for: .bibleViewSwipeMode), "CHAPTER")
-        XCTAssertEqual(AppPreferenceRegistry.stringDefault(for: .notesContentType), "HTML")
-        XCTAssertEqual(AppPreferenceRegistry.intDefault(for: .fontSizeMultiplier), 100)
-        XCTAssertEqual(AppPreferenceRegistry.intRange(for: .fontSizeMultiplier), 10...500)
-        XCTAssertEqual(AppPreferenceRegistry.boolDefault(for: .openLinksInSpecialWindowPref), true)
-        XCTAssertEqual(AppPreferenceRegistry.boolDefault(for: .enableBluetoothPref), true)
-        XCTAssertEqual(AppPreferenceRegistry.boolDefault(for: .bookGridLeftToRight), false)
-        XCTAssertEqual(AppPreferenceRegistry.boolDefault(for: .bookGridGroupByCategory), false)
-        XCTAssertEqual(AppPreferenceRegistry.boolDefault(for: .bookGridShowLongName), false)
-        XCTAssertEqual(AppPreferenceRegistry.boolDefault(for: .bookGridShowProgress), true)
-    }
-
-    func testNotesContentTypeNormalizerUsesAndroidSupportedValues() {
-        XCTAssertEqual(AppPreferenceValueNormalizer.notesContentType("HTML"), "HTML")
-        XCTAssertEqual(AppPreferenceValueNormalizer.notesContentType("MARKDOWN"), "MARKDOWN")
-        XCTAssertEqual(AppPreferenceValueNormalizer.notesContentType("markdown"), "HTML")
-        XCTAssertEqual(AppPreferenceValueNormalizer.notesContentType("PLAINTEXT"), "HTML")
-        XCTAssertEqual(AppPreferenceValueNormalizer.notesContentType(""), "HTML")
-    }
-
-    func testActionPreferencesUseActionShape() {
-        let actionKeys: [AppPreferenceKey] = [
-            .discreteHelp,
-            .openLinks,
-            .crashApp,
-        ]
-
-        for key in actionKeys {
-            let definition = AppPreferenceRegistry.definition(for: key)
-            if case .action = definition.storage {
-                // expected
-            } else {
-                XCTFail("Expected .action storage for \(key.rawValue)")
-            }
-            if case .action = definition.valueType {
-                // expected
-            } else {
-                XCTFail("Expected .action valueType for \(key.rawValue)")
-            }
-            XCTAssertNil(definition.defaultValue)
-        }
-    }
-
-    func testApplicationPreferencesResetContractExcludesActionRowsAndIncludesVisibleParityKeys() {
-        let resetKeys = AppPreferenceRegistry.applicationPreferencesResetKeys
-
-        XCTAssertTrue(resetKeys.contains(.navigateToVersePref))
-        XCTAssertTrue(resetKeys.contains(.volumeKeysScroll))
-        XCTAssertTrue(resetKeys.contains(.localePref))
-        XCTAssertTrue(resetKeys.contains(.calculatorPin))
-        XCTAssertTrue(resetKeys.contains(.experimentalFeatures))
-        XCTAssertFalse(resetKeys.contains(.discreteHelp))
-        XCTAssertFalse(resetKeys.contains(.openLinks))
-        XCTAssertFalse(resetKeys.contains(.crashApp))
-        XCTAssertEqual(Set(resetKeys).count, resetKeys.count)
-    }
-
-    func testSettingsStoreResetApplicationPreferencesRestoresRegistryDefaults() throws {
-        let userDefaultsKeys: [AppPreferenceKey] = [.localePref, .showCalculator, .calculatorPin, .discreteMode]
-        userDefaultsKeys.forEach { UserDefaults.standard.removeObject(forKey: $0.rawValue) }
-        defer {
-            userDefaultsKeys.forEach { UserDefaults.standard.removeObject(forKey: $0.rawValue) }
-        }
-
-        let settingsStore = try makeInMemorySettingsStore()
-        settingsStore.setBool(.navigateToVersePref, value: true)
-        settingsStore.setString(.toolbarButtonActions, value: "swap-menu")
-        settingsStore.setString(.notesContentType, value: "MARKDOWN")
-        settingsStore.setInt(.fontSizeMultiplier, value: 180)
-        settingsStore.setStringSet(.experimentalFeatures, values: ["feature_b", "feature_a"])
-        settingsStore.setString(.localePref, value: "fi")
-        settingsStore.setBool(.showCalculator, value: true)
-        settingsStore.setString(.calculatorPin, value: "9999")
-        settingsStore.setBool(.discreteMode, value: true)
-        settingsStore.setString(SettingsStore.globalTextDisplaySettingsKey, value: #"{"fontSize":22}"#)
-
-        settingsStore.resetApplicationPreferences()
-
-        XCTAssertEqual(settingsStore.getBool(.navigateToVersePref), false)
-        XCTAssertEqual(settingsStore.getString(.toolbarButtonActions), "default")
-        XCTAssertEqual(settingsStore.getString(.notesContentType), "HTML")
-        XCTAssertEqual(settingsStore.getInt(.fontSizeMultiplier), 100)
-        XCTAssertEqual(settingsStore.getStringSet(.experimentalFeatures), [])
-        XCTAssertEqual(settingsStore.getString(.localePref), "")
-        XCTAssertEqual(settingsStore.getBool(.showCalculator), false)
-        XCTAssertEqual(settingsStore.getString(.calculatorPin), "1234")
-        XCTAssertEqual(settingsStore.getBool(.discreteMode), false)
-        XCTAssertNotNil(settingsStore.getString(SettingsStore.globalTextDisplaySettingsKey))
-    }
-
-    func testSettingsSearchMatcherRequiresAllNormalizedTermsAcrossEntryText() {
-        let entry = AndBibleSettingsSearchEntry(
-            identifier: "settingsReadingProgressLink",
-            title: "Reading Progress Settings",
-            summary: "Configure automatic reading tracking",
-            detail: "Memorization",
-            keywords: ["features", "progress"]
-        )
-
-        XCTAssertTrue(AndBibleSettingsSearchMatcher.matches(query: "", entry: entry))
-        XCTAssertTrue(AndBibleSettingsSearchMatcher.matches(query: "reading tracking", entry: entry))
-        XCTAssertTrue(AndBibleSettingsSearchMatcher.matches(query: "FEATURES progress", entry: entry))
-        XCTAssertFalse(AndBibleSettingsSearchMatcher.matches(query: "sync tracking", entry: entry))
-    }
-
-    func testSettingsSearchMatcherFiltersExactRenderedRowsWithinMatchingSection() {
-        let entries = [
-            AndBibleSettingsSearchEntry(
-                identifier: "monochrome_mode",
-                title: "Black & white mode",
-                summary: "Use application in monochrome mode"
-            ),
-            AndBibleSettingsSearchEntry(
-                identifier: "disable_animations",
-                title: "Disable animations",
-                summary: "Disable smooth scrolling animations"
-            ),
-        ]
-
-        XCTAssertTrue(entries.contains { AndBibleSettingsSearchMatcher.matches(query: "monochrome", entry: $0) })
-        XCTAssertTrue(
-            AndBibleSettingsSearchMatcher.matchesIdentifier(
-                "monochrome_mode",
-                query: "monochrome",
-                entries: entries
-            )
-        )
-        XCTAssertFalse(
-            AndBibleSettingsSearchMatcher.matchesIdentifier(
-                "disable_animations",
-                query: "monochrome",
-                entries: entries
-            )
-        )
-        XCTAssertFalse(
-            AndBibleSettingsSearchMatcher.matchesIdentifier(
-                "missing_row",
-                query: "monochrome",
-                entries: entries
-            )
-        )
-        XCTAssertTrue(
-            AndBibleSettingsSearchMatcher.matchesIdentifier(
-                "missing_row",
-                query: "",
-                entries: entries
-            )
-        )
-    }
-
     func testTextDisplayAppDefaultsUseAndroidHiddenLinksStrongsMode() {
         XCTAssertEqual(TextDisplaySettings.appDefaults.strongsMode, 0)
     }
