+++ docs/test-architecture-migration-plan.md
@@ -152,6 +152,7 @@ Blocker: `AndBibleTestSupport.swift` is a 2,233-line `extension AndBibleTests` e
 - `AndBibleTests+Downloads` has been split by owning module: Downloads browser presentation, Android filter/default-language, auto-refresh, cancellation, and localized error-copy contracts moved to `BibleUITests`; queued search-index deletion and MyBible reader payload readability moved to `BibleCoreTests`; SWORD/MyBible repository refresh, install layout, rollback, package fallback, ZIP import, and cancellation contracts moved to `SwordKitTests`. The duplicated restored-MyBible `SwordManager.installedModules()` cases were not re-migrated because equivalent package coverage already lives in `SwordManagerTests`.
 - `AndBibleTests+Bookmarks` has been split by owning module: BookmarkService note/label/paragraph-break/StudyPad persistence contracts moved to `BibleCoreTests`, while bookmark-list reference formatting, reader bookmark/note/StudyPad bridge payloads, label-assignment routing, and reader accessibility snapshots moved to `BibleUITests`. JSword range parity still uses the shared bundled-SWORD fixture from app resources; moving that fixture into package test resources remains a follow-up cleanup, not a reason to keep these tests in the app-host bundle.
 - `AndBibleTests+AppAndReader` boundary work has started: iPadOS windowing-control policy tests moved to `BibleUITests`, `BibleWebView` platform bootstrap tests moved to `BibleViewTests`, CI now runs the `BibleViewTests` package scheme, and `testContentViewDoesNotContainLegacyRootSidebarShell` became a repo-standards `source-guards` check. `testApplicationDelegateSceneConfigurationUsesWindowSceneDelegate` remains app-hosted because it validates app delegate scene wiring.
+- `AndBibleTests+AppAndReader` settings slice has been split by owning module: application preference registry/default/normalization/reset contracts moved to `BibleCoreTests`, while settings-search matcher presentation contracts moved to `BibleUITests`.
 - Batches for the genuinely BibleUI-behavior remainder: ReaderNavigation, Bookmarks/Strongs, Window/Settings, view-model parts of `+AppAndReader`.
 - **Split `+AppAndReader` deliberately**: `sceneConfiguration` test stays app-hosted; the `ContentView` legacy root-sidebar source scan is now a repo-standards `source-guards` check; the remaining package-owned tests should move to the lowest owning package target.
 - Run BibleUITests via `xcodebuild test -scheme BibleUITests -destination 'platform=iOS Simulator,...'` against the **committed package test scheme from Phase 0** (no app host).
